Feita algumas mudanças no arquivo tables.sql e app/routes/users.py . Foi criado um novo arquivo na pasta dao para separar os comandos do banco da rota de usuários
